backend/pyproject.toml
[tool.poetry]
name = "rag-experiments-backend"
version = "0.1.0"
description = ""
authors = ["Isabelle"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
llama-index = "0.10.50"
llama-index-core = "0.10.50"
llama-index-vector-stores-supabase = "^0.1.5"
fastapi = "^0.112.0"
dependency-injector = "^4.41.0"
pyjwt = "^2.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
package-mode = false
